feat: move under `/profile`
