+++ src/app/(public)/profile/[slug]/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <div>Loading profile...</div>;
+}
